internal/worker: fix an issue where worker won't restart
Fix a case where the worker refuses to restart if it thinks
(incorrectly) that a previous update has not yet finished. (Because
the previous update errored).

Do this by populating the EndedAt field of the UpdateRecord in
cveUpdater.update on error as well as success.

Change-Id: Ibc98f23b51da21397612479d539dfefe906acb8c
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/589996
Reviewed-by: Damien Neil <dneil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Tatiana Bradley <tatianabradley@google.com>