cmd/cve: add command cve reject
Adds a command that allows the user (authenticated as a CNA) to
"reject" a CVE that will never be published.

This is intended for CVEs that were reserved but will never be assigned
to a vulnerability because, e.g, the year is not current.

Change-Id: Id60a0e5417d43e791ada898ff83bcef2563c2322
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/592435
Auto-Submit: Tatiana Bradley <tatianabradley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>