internal/genericosv: work around bug in Severity unmarshal
There is an unknown bug causing unmarshal for OSV Severity field
to not work.

We don't use this field, so to work around this issue for now,
simply ignore the Severity field.

To do this, the dependency on osv-scanner was removed and the
relevant files were copied and modified as needed.

Change-Id: I956ea5d2c9c19f2992e6a1c9b723cea35f5e92d6
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/593817
Auto-Submit: Tatiana Bradley <tatianabradley@google.com>
Reviewed-by: Damien Neil <dneil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>