internal/osvutils: don't require references in OSV
Individual lint checks cover the cases in which certain references
are required, so the check for having references is not needed.

(For example, standard library reports have certain requirements,
and reports with no description must list an advisory).

In some limited cases, such as when a report is the first
public reference to a vulnerability, it can be appropriate to
have no references at all.

Change-Id: Ide58fd1422c06eb44b83ae6c3f74b49e3fdf5b48
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/602136
Reviewed-by: Damien Neil <dneil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Tatiana Bradley <tatianabradley@google.com>