internal/triage: add lots of tests for triage
To prep for a change in the triage algorithm, add 100 test cases
for real CVEs (75 Go and 25 not Go) and their triage results
with the current algorithm. (The current algorithm is copied
into a new function, AffectsGo, with no changes).

Note that the current algorithm doesn't always get it right, so
each test output file marks whether the result is correct or not.

This is OK; CVEs don't explicitly mark themselves as affecting
Go or not so we can't expect to always correctly categorize them.
However, we'd like to make sure changing the algorithm doesn't
reduce correctness.

Change-Id: I00ce8519b997d8605ab2e4693980f256d8eab7d0
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/602597
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>