data/reports: mark GO-2025-3442 as REVIEWED
GO-2025-3442 seems to have been reviewed in the past: information such
as symbols are available, despite vulnreport not being able to infer it
automatically.

The unexpected status of GO-2025-3442 is currently causing a panic while
running vulnreport for golang/vulndb#4025. This is because while
processing golang/vulndb#4025, vulnreport tries to analyze past reports
for the same module, and NEEDS_REVIEW status is not considered a valid
status for this purpose.

As such, let us update the review status of GO-2025-3442 to REVIEWED.

Change-Id: Iba0c73497b0ffc7de45271251ea414d6954468b0
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/721084
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Markus Kusano <kusano@google.com>
Reviewed-by: Ethan Lee <ethanalee@google.com>
Commit-Queue: Markus Kusano <kusano@google.com>